Enhancement to BGP test cases.
